+++ .gitignore
@@ -1,0 +1,3 @@
+config.json
+*.pyc
+weibo/
