+++ config_sample.json
@@ -1,0 +1,18 @@
+{
+    "filter": 1,
+    "since_date": "2018-01-01",
+    "write_mode": "txt",
+    "pic_download": 1,
+    "video_download": 1,
+    "cookie": "your cookie",
+    "mysql_config": {
+        "host": "localhost",
+        "port": 3306,
+        "user": "root",
+        "password": "123456",
+        "charset": "utf8mb4"
+    },
+    "user_id_list": [
+        "1669879400"
+    ]
+}
